+++ test/printjs/printjs_example.compiled.mjs
@@ -66,11 +66,10 @@ export default async (arg$0) => {
                   if (!(errors$0.contents.length === 0)) {
                     errors$0.contents += "\n\n";
                   }
-                  errors$0.contents += "File \"";
+                  errors$0.contents += "Error while rendering \"";
                   errors$0.contents += "printjs_example.acutis";
                   errors$0.contents +=
-                    "\"\n\
-Render error.\n\
+                    "\".\n\
 The data supplied does not match this template's interface.\n\
 ";
                   errors$0.contents += "Path:\n<input>";